[09/01] feat: Support decimal numbers
A problem is that numbers with more than 6 digits get printed in scientific notation.
Printing in fixed point with 0 precision is not a solution since `5 2 /` prints `2` and then `2 *` prints `5` again.

+++ cpp09/ex01/src/RPN.hpp
@@ -18,12 +18,12 @@ public:
 	RPN& operator=(const RPN& other);
 	~RPN();
 
-	ft::Expected<long, std::string> calculate(const std::string& input);
-	ft::Expected<long, std::string> result();
+	ft::Expected<long double, std::string> calculate(const std::string& input);
+	ft::Expected<long double, std::string> result();
 
 private:
 	void _push_operator(Token op);
 	void _push_operand(const std::string& word);
 
-	std::stack<long> _stack;
+	std::stack<long double> _stack;
 };
